Move publishing to a module (#26)
This change puts all the publishing machinery into a `publish` module,
and includes this module if the corresponding feature is enabled (which
it is by default).

This will help to organize the crate better, especially as the consumer
half of the API will be added in the coming weeks.

Along with the module move, the `Message` type was renamed to
`EncodableMessage`, to make it clear that this is for encoding. A
corresponding DecodableMessage trait will be included with the consumer
API, using a separate trait to enable encoding borrowed types and
decoding into owned types.
