--- src/publishers.rs
@@ -1,12 +1,0 @@
-//! [`Publisher`](crate::Publisher) implementations.
-
-mod null;
-pub use null::*;
-
-mod mock;
-pub use mock::*;
-
-#[cfg(feature = "google")]
-mod googlepubsub;
-#[cfg(feature = "google")]
-pub use googlepubsub::*;
